+++ src/main/java/org/example/repository/ProjectRepo.java
@@ -1,5 +1,35 @@
 package org.example.repository;
 
-public class ProjectRepo {
+import org.example.entity.Project;
 
+import java.util.HashMap;
+import java.util.Map;
+
+public class ProjectRepo {
+    public static Map<String, Project> projectMap = new HashMap<>();
+
+
+    public void findAll() {
+
+    }
+
+    public void findOne() {
+
+    }
+
+    public void persist() {
+
+    }
+
+    public void merge() {
+
+    }
+
+    public void remove() {
+
+    }
+
+    public void removeAll() {
+
+    }
 }
