+++ src/main/java/org/example/repository/TaskRepo.java
@@ -1,4 +1,35 @@
 package org.example.repository;
 
+import org.example.entity.Task;
+
+import java.util.HashMap;
+import java.util.Map;
+
 public class TaskRepo {
+    public static Map<String, Task> taskMap = new HashMap<>();
+
+
+    public void findAll() {
+
+    }
+
+    public void findOne() {
+
+    }
+
+    public void persist() {
+
+    }
+
+    public void merge() {
+
+    }
+
+    public void remove() {
+
+    }
+
+    public void removeAll() {
+
+    }
 }
